Hopefully fixed centering map issue.
Before when the paged refreshed and it was not active the element didn't have a width so the image would be left aligned to the center instead of the center being in the center.  I'm sure there is a better way to do this, however, I'm not going to figure it out as we may change the map to a live map in the near future.